requirements.txt
runpod>=1.7.0
huggingface_hub>=0.24.0
transformers>=4.44.0
diffusers>=0.30.0
accelerate>=0.33.0
safetensors>=0.4.5
einops>=0.8.0
omegaconf>=2.3.0
sentencepiece>=0.2.0
pillow>=10.4.0
numpy<2.0
opencv-python-headless>=4.10.0
imageio>=2.35.0
imageio-ffmpeg>=0.5.1
av>=12.3.0
requests>=2.32.0
mediapy>=1.2.0
rotary-embedding-torch>=0.6.5
